justdata/apps/loantrends/requirements.txt
# LoanTrends Dependencies
# Python 3.8 or higher required

# Web Framework
flask>=2.3.0
Werkzeug>=2.3.0

# Production Server (for Render)
gunicorn>=20.1.0

# Data Processing
pandas>=1.5.0
numpy>=1.21.0

# Excel Generation
openpyxl>=3.0.0

# Environment Variables
python-dotenv>=1.0.0

# HTTP Requests
requests>=2.28.0

# AI/LLM providers
anthropic>=0.34.0

# Logging
structlog>=23.0.0




